Create Orange provider template
