+++ resources/lib/provider_templates/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# pylint: disable=missing-module-docstring
+from .orange import OrangeTemplate
